% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_pa.R
\name{ggplot_pa}
\alias{ggplot_pa}
\title{Parallel analysis plot via ggplot2}
\usage{
ggplot_pa(x, bw = TRUE, ...)
}
\arguments{
\item{x}{An object of class parallel returned from the fa.parallel function}

\item{bw}{Do you want a black and white theme? Accepts TRUE or FALSE.}
}
\description{
\code{ggplot_pa} returns a parallel analysis plot using ggplot2
}
\examples{
pa.cog <- fa.parallel(cognition, fm = "ml")
ggplot_pa(pa.cog)

}

