% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_scree.R
\name{ggplot_scree}
\alias{ggplot_scree}
\title{Parallel analysis plot via ggplot2}
\usage{
ggplot_scree(x, bw = TRUE, ...)
}
\arguments{
\item{x}{A correlation matrix}

\item{bw}{Do you want the black and white theme? Accepts TRUE or FALSE.}
}
\description{
\code{ggplot_scree} returns a scree plot using ggplot2
}
\examples{
interest.cor <- cor(interest)
ggplot_scree(interest.cor)
}

