% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latticeScree.R
\name{latticeScree}
\alias{latticeScree}
\title{Scree plot via lattice}
\usage{
latticeScree(x, ...)
}
\arguments{
\item{x}{A correlation matrix}
}
\description{
\code{latticeScree} returns a scree plot using the lattice package
}
\examples{
interest.cor <- cor(interest)
latticescree(interest.cor)
}

